chore: move password util to separate pkg
